:hammer: Update skip event handlers in dom-change event
Previously, the touch-track event handlers were attached to the first track when the currentTrack inside the app-state changed. This forced us to use setTimeout to defer the dom-querying into the next tick, because the dom is only updated after all property observers have fired. Now we use the dom-change event, which is fired every time dom-repeat changes it's light dom. This is more reliable and cleaner.